requirements.txt
# Make sure that:
# 1) Python version is 3.9.21
# 2) CUDA version is 12.2

# Install PyTorch by:
# pip install torch==2.4.1 torchvision==0.19.1 torchaudio==2.4.1 --index-url https://download.pytorch.org/whl/cu121
# pip install torch-scatter -f https://data.pyg.org/whl/torch-2.4.1+cu121.html

# Then do pip install this file

auto_mix_prep==0.2.0
autorootcwd==1.0.1
hydra-core==1.3.2
lightning==2.4.0
lxml==5.3.0
mlflow==2.18.0
nltk==3.9.1
numpy==1.26.4
omegaconf==2.3.0
pandas==2.2.3
pytorch_lightning==2.4.0
PyYAML==6.0.2
tqdm==4.66.5
transformers==4.46.1
Unidecode==1.3.8
